ch02/requirements.txt
numpy
pillow
tensorflow
flask 